+++ app/services/user/UserService.scala
@@ -13,7 +13,7 @@ trait UserServiceComponent {
 
     def updateUser(user: User): Boolean
 
-    def tryFindById(id: Long): Option[User]
+    def tryFindById(id: Int): Option[User]
 
     def delete(id: Long)
 
@@ -36,7 +36,7 @@ trait UserServiceComponentImpl extends UserServiceComponent {
       userRepository.updateUser(user)
     }
 
-    override def tryFindById(id: Long): Option[User] = {
+    override def tryFindById(id: Int): Option[User] = {
       userRepository.tryFindById(id)
     }
